deps: add logic to moksha-wallet dependencies
While attempting to compile for esp32 boards using esp-rs crates, moksha-wallet couldn't
be depended on because of various compile errors. When compiling for esp, you are not targeting
wasm. Irconically, the wasm dependencies mostly compiled for esp. This commit adds logic so
espcash can depend directly on moksha to more easily implement an ecash clinet for embedded devices.